Add memspace "highest bandwidth"
This memspace contains an aggregated list of NUMA nodes identified as best
targets after selecting each NUMA node as the initiator.
Querying the bandwidth value requires HMAT support on the platform,
calling umfMemspaceHighestBandwidthGet() will return NULL if it's not
supported.
Introduce UMF_MEMSPACE_HIGHEST_BANDWIDTH env var. This environment variable
is used to forcefully assign NUMA nodes to the memspace "highest bandwidth".
The ordering of memory targets within this memspace will correspond to the
sequence of NUMA nodes specified in the environment variable.